+++ server/admin.go
@@ -68,7 +68,6 @@ func NewAdmin(controller *Controller) *Admin {
 }
 
 func (admin *Admin) BroadcastConfig() {
-
 	if b, err := json.Marshal(admin.GetConfig()); err == nil {
 		for conn := range admin.Conns {
 			conn.WriteMessage(websocket.TextMessage, b)
@@ -76,7 +75,7 @@ func (admin *Admin) BroadcastConfig() {
 	}
 }
 
-func (admin *Admin) ChangePassword(currentPassword string, newPassword string) error {
+func (admin *Admin) ChangePassword(currentPassword interface{}, newPassword string) error {
 	var (
 		err  error
 		hash []byte
@@ -86,8 +85,11 @@ func (admin *Admin) ChangePassword(currentPassword string, newPassword string) e
 		return errors.New("newPassword is empty")
 	}
 
-	if err = bcrypt.CompareHashAndPassword([]byte(admin.Controller.Options.adminPassword), []byte(currentPassword)); err != nil {
-		return err
+	switch v := currentPassword.(type) {
+	case string:
+		if err = bcrypt.CompareHashAndPassword([]byte(admin.Controller.Options.adminPassword), []byte(v)); err != nil {
+			return err
+		}
 	}
 
 	if hash, err = bcrypt.GenerateFromPassword([]byte(newPassword), bcrypt.DefaultCost); err != nil {
@@ -141,13 +143,6 @@ func (admin *Admin) ConfigHandler(w http.ResponseWriter, r *http.Request) {
 		}()
 
 	} else {
-		admin.Controller.IngestLock()
-		admin.mutex.Lock()
-		defer func() {
-			admin.mutex.Unlock()
-			admin.Controller.IngestUnlock()
-		}()
-
 		logError := func(err error) {
 			admin.Controller.Logs.LogEvent(LogLevelError, fmt.Sprintf("admin.confighandler.put: %s", err.Error()))
 		}
@@ -169,6 +164,9 @@ func (admin *Admin) ConfigHandler(w http.ResponseWriter, r *http.Request) {
 				w.WriteHeader(http.StatusBadRequest)
 				return
 			}
+
+			admin.Controller.IngestLock()
+			admin.mutex.Lock()
 
 			admin.Controller.Dirwatches.Stop()
 
@@ -279,6 +277,9 @@ func (admin *Admin) ConfigHandler(w http.ResponseWriter, r *http.Request) {
 				}
 			}
 
+			admin.mutex.Unlock()
+			admin.Controller.IngestUnlock()
+
 			admin.Controller.EmitConfig()
 			admin.Controller.Dirwatches.Start(admin.Controller)
 
@@ -349,13 +350,14 @@ func (admin *Admin) LogsHandler(w http.ResponseWriter, r *http.Request) {
 
 		r, err := admin.Controller.Logs.Search(&logOptions, admin.Controller.Database)
 		if err != nil {
-			fmt.Println(err)
+			admin.Controller.Logs.LogEvent(LogLevelError, err.Error())
 			w.WriteHeader(http.StatusExpectationFailed)
 			return
 		}
 
 		b, err := json.Marshal(r)
 		if err != nil {
+			admin.Controller.Logs.LogEvent(LogLevelError, err.Error())
 			w.WriteHeader(http.StatusExpectationFailed)
 			return
 		}
@@ -486,7 +488,7 @@ func (admin *Admin) PasswordHandler(w http.ResponseWriter, r *http.Request) {
 	case http.MethodPost:
 		var (
 			b               []byte
-			currentPassword string
+			currentPassword interface{}
 			newPassword     string
 		)
 
@@ -510,9 +512,6 @@ func (admin *Admin) PasswordHandler(w http.ResponseWriter, r *http.Request) {
 		switch v := m["currentPassword"].(type) {
 		case string:
 			currentPassword = v
-		default:
-			w.WriteHeader(http.StatusBadRequest)
-			return
 		}
 
 		switch v := m["newPassword"].(type) {
@@ -524,14 +523,12 @@ func (admin *Admin) PasswordHandler(w http.ResponseWriter, r *http.Request) {
 		}
 
 		if err = admin.ChangePassword(currentPassword, newPassword); err != nil {
-			logError(err)
+			logError(errors.New("unable to change admin password, current password is invalid"))
 			w.WriteHeader(http.StatusExpectationFailed)
 			return
 		}
 
-		if b, err = json.Marshal(map[string]interface{}{
-			"passwordNeedChange": admin.Controller.Options.adminPasswordNeedChange,
-		}); err == nil {
+		if b, err = json.Marshal(map[string]interface{}{"passwordNeedChange": admin.Controller.Options.adminPasswordNeedChange}); err == nil {
 			w.Write(b)
 		} else {
 			w.WriteHeader(http.StatusExpectationFailed)
@@ -599,6 +596,86 @@ func (admin *Admin) Start() error {
 	}()
 
 	return nil
+}
+
+func (admin *Admin) UserAddHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case http.MethodPost:
+		logError := func(err error) {
+			admin.Controller.Logs.LogEvent(LogLevelError, fmt.Sprintf("admin.useraddhandler.post: %s", err.Error()))
+		}
+
+		t := admin.GetAuthorization(r)
+		if !admin.ValidateToken(t) {
+			w.WriteHeader(http.StatusUnauthorized)
+			return
+		}
+
+		m := map[string]interface{}{}
+		err := json.NewDecoder(r.Body).Decode(&m)
+		if err != nil {
+			w.WriteHeader(http.StatusBadRequest)
+			return
+		}
+
+		admin.Controller.Accesses.Add(NewAccess().FromMap(m))
+
+		if err := admin.Controller.Accesses.Write(admin.Controller.Database); err == nil {
+			if err := admin.Controller.Accesses.Read(admin.Controller.Database); err == nil {
+				admin.BroadcastConfig()
+				w.WriteHeader(http.StatusOK)
+			} else {
+				logError(err)
+				w.WriteHeader(http.StatusExpectationFailed)
+			}
+		} else {
+			logError(err)
+			w.WriteHeader(http.StatusExpectationFailed)
+		}
+
+	default:
+		w.WriteHeader(http.StatusMethodNotAllowed)
+	}
+}
+
+func (admin *Admin) UserRemoveHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case http.MethodPost:
+		logError := func(err error) {
+			admin.Controller.Logs.LogEvent(LogLevelError, fmt.Sprintf("admin.userremovehandler.post: %s", err.Error()))
+		}
+
+		t := admin.GetAuthorization(r)
+		if !admin.ValidateToken(t) {
+			w.WriteHeader(http.StatusUnauthorized)
+			return
+		}
+
+		m := map[string]interface{}{}
+		err := json.NewDecoder(r.Body).Decode(&m)
+		if err != nil {
+			w.WriteHeader(http.StatusBadRequest)
+			return
+		}
+
+		if _, ok := admin.Controller.Accesses.Remove(NewAccess().FromMap(m)); ok {
+			if err := admin.Controller.Accesses.Write(admin.Controller.Database); err == nil {
+				if err := admin.Controller.Accesses.Read(admin.Controller.Database); err == nil {
+					admin.BroadcastConfig()
+					w.WriteHeader(http.StatusOK)
+				} else {
+					logError(err)
+					w.WriteHeader(http.StatusExpectationFailed)
+				}
+			} else {
+				logError(err)
+				w.WriteHeader(http.StatusExpectationFailed)
+			}
+		}
+
+	default:
+		w.WriteHeader(http.StatusMethodNotAllowed)
+	}
 }
 
 func (admin *Admin) ValidateToken(sToken string) bool {